async-main
Removed the coreclr loading workarounds and utilized an entry point in `NodeHostEnvironment` to call the app.
This removes the need to load the native libraries from NodeHostEnvironment and disposing the host environment manually.
This requires an async main entry point in the app to keep the host environment alive.

Some minor cleanups and improvements are also included.