Adapted passing of .Net or native memory to JS to new ArrayBuffer backing store implementation
We have to ensure that we do not pass the same memory multiple times to JS.
This is done by using a `map` of memory already passed to JS.

This lead to improving the `NativeMemory` API and making it more safe. Published native memory is pooled now and only reference counted handles to it are accessed.